requirements.txt
# Core
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# CrewAI
crewai==0.28.8
crewai-tools==0.1.6

# AI & Tools
google-generativeai==0.3.2
pydantic==2.5.0

# Utilities
requests==2.31.0
markdown==3.5.1